refactor: opacity hook으로 분리 후 적용
