+++ src/pages/home/subPages/dashboard/components/MemoSectionList/MemoSectionList.tsx
@@ -1,9 +1,10 @@
 import { Tag, TagRelation } from 'pages/home/subPages/interfaces';
 import { MemoSection } from './MemoSection';
 import { v4 as uuid_v4 } from 'uuid';
-import { useContext, useEffect, useRef, useState } from 'react';
+import { useContext, useEffect, useRef } from 'react';
 import {
   useHorizontalScroll,
+  useHorizontalScrollOpacity,
   useIntersectionObserver,
 } from 'pages/home/subPages/hooks';
 import { SortOption } from 'pages/home/subPages/types';
@@ -24,31 +25,11 @@ const MemoSectionList = ({
 }: MemoSectionListProps) => {
   const { selectedTag } = useContext(TagContext);
   const observerRef = useRef<HTMLDivElement | null>(null);
-  const scrollRef = useRef<HTMLDivElement | null>(null);
+  const { scrollRef, scrollOpacity } = useHorizontalScrollOpacity();
 
   const { onDragStart, onDragMove, onDragEnd } = useHorizontalScroll({ scrollRef });
 
   const hasNoSection = () => tagRelations.length === 0;
-
-  const [scrollOpacity, setScrollOpacity] = useState({ left: 1, right: 1 });
-
-  const updateScrollOpacity = () => {
-    if (scrollRef.current) {
-      const { scrollLeft, scrollWidth, clientWidth } = scrollRef.current;
-      const leftOpacity = scrollLeft > 0 ? 0.1 : 1;
-      const rightOpacity = scrollLeft + clientWidth < scrollWidth ? 0.1 : 1;
-      setScrollOpacity({ left: leftOpacity, right: rightOpacity });
-    }
-  };
-
-  useEffect(() => {
-    const scrollElement = scrollRef.current;
-    if (scrollElement) {
-      updateScrollOpacity();
-      scrollElement.addEventListener('scroll', updateScrollOpacity);
-      return () => scrollElement.removeEventListener('scroll', updateScrollOpacity);
-    }
-  }, []);
 
   useIntersectionObserver(observerRef, {
     callback: (entries) => {
