+++ src/pages/home/subPages/hooks/index.ts
@@ -2,3 +2,5 @@ export { default as usePressEnterFetch } from './usePressEnterFetch';
 export { default as useIntersectionObserver } from './useIntersectionObserver';
 export { default as useHorizontalScroll } from './useHorizontalScroll';
 export { default as useImageList } from './useImageList';
+export { default as useVerticalScrollOpacity } from './useVerticalScrollOpacity';
+export { default as useHorizontalScrollOpacity } from './useHorizontalScrollOpacity';
